// types/business.ts
export type BusinessForm = {
  business_name: string;
  tagline: string;

  address: string;

  phone: string;
  email: string;
  instagram: string;

  opening_hours: string;

  maps_url: string;
  maps_embed: string;

  latitude: string;
  longitude: string;

  price_range: string;

  logo: string;
};